requirements.txt
Flask == 3.0.3
icecream == 2.1.3
mysql-connector-python == 9.0.0
flask-session == 0.8.0
gspread == 6.2.1
oauth2client == 4.1.3
requests == 2.32.5
mysql-connector-python==8.1.0
